update parameter type

+++ squirrel-foundation/src/main/java/org/squirrelframework/foundation/fsm/ImmutableTimedState.java
@@ -2,9 +2,9 @@ package org.squirrelframework.foundation.fsm;
 
 public interface ImmutableTimedState<T extends StateMachine<T, S, E, C>, S, E, C> extends ImmutableState<T, S, E, C> {
     
-    Integer getTimeInterval();
+    long getTimeInterval();
     
-    Integer getInitialDelay();
+    long getInitialDelay();
     
     E getAutoFireEvent();
