+++ squirrel-foundation/src/test/java/org/squirrelframework/foundation/fsm/snake/GameConfigure.java
@@ -10,5 +10,5 @@ public class GameConfigure {
 	
 	public static final int CELL_SIZE = 16;
 	
-	public static final int FRAME_TIME = 150; // ms
+	public static final long FRAME_TIME = 150; // ms
 }
